remove old system
