+++ src/cli/helpers/index.ts
@@ -1,7 +1,3 @@
 export { HelpError, unknownCommand } from './Help'
-export {
-  Command,
-  Commands,
-  Dictionary,
-} from './types'
+export { Command, Commands, Dictionary } from './types'
 export { arg, format, isError } from './utils'
